linking the uart code to the cmake list file
